Refactor ListGroup's file & folder structure
